Changes Store#increment to return an Int64

+++ src/defense/memory_store.cr
@@ -1,13 +1,13 @@
 module Defense
   class MemoryStore < Store
     def initialize
-      @data = Hash(String, Int32).new # how to represent the time?!
+      @data = Hash(String, Int64).new
     end
 
-    def increment(key : String, expires_in : Int32) : Int32
+    def increment(key : String, expires_in : Int32) : Int64
       # check that it's not expired
       # if expired => reset
-      @data[key] = (@data[key]? || 0) + 1
+      @data[key] = (@data[key]? || 0i64) + 1i64
     end
 
     def reset
