+++ src/defense/store.cr
@@ -1,6 +1,6 @@
 module Defense
   abstract class Store
-    abstract def increment(key : String, expires_in : Int32) : Int32
+    abstract def increment(key : String, expires_in : Int32) : Int64
     abstract def reset
   end
 end
